Extract date to fields package
