+++ packages/fields/src/fields/index.ts
@@ -6,3 +6,4 @@ export { default as parentField } from './parent';
 export { default as passwordField } from './password';
 export { default as statusField } from './status';
 export { default as commentStatusField } from './comment-status';
+export { default as dateField } from './date';
